Moved utils to search folder from views
