+++ firstvoices/backend/models/dictionary.py
@@ -4,7 +4,7 @@ from django.utils.translation import gettext as _
 
 from backend.permissions import predicates
 from backend.search.indices.dictionary_entry_document import DictionaryEntryDocument
-from backend.search.utils import elasticsearch_running
+from backend.search.utils.search_utils import elasticsearch_running
 from backend.utils.character_utils import clean_input
 
 from .base import (
